To allow old copies of inskeys to still work on the installer image we switch back to the old BKLXXX template for the config data.
Without this the 4.0-beta[123] releases were unable to upgrade
to this release. 

bk: 44b78b1e1K-EO0izIyuFilLHM0LdbQ

+++ src/upgrade.c
@@ -252,12 +252,13 @@ next:				freeLines(data, free);
 		goto out;
 	}
 	fclose(f);
-	rc = inskeys(data[1], licf);
-	unlink(licf);
-	free(licf);
-	if (rc) goto out;
-	rc = 2;
-
+	unless (getenv("_BK_UPGRADE_NOINSKEYS")) {
+		rc = inskeys(data[1], licf);
+		unlink(licf);
+		free(licf);
+		if (rc) goto out;
+		rc = 2;
+	}
 	chmod(data[1], 0500);
 
 	if (fetchonly) {
